Refactor to avoid "common".
